+++ react/components/Dropdown/index.js
@@ -49,10 +49,9 @@ class Dropdown extends Component {
     const placeholderValue = placeholder || label || helpText || ''
 
     if (placeholderValue === '' && !this.sentPlaceholderWarning) {
+      const identificationText = this.getDropdownIdentification()
       console.warn(
-        `The following dropdown has a placeholder option, but no placeholder text. Please use at least one of these props: placeholder, label, or helpText.
-
-${this.getDropdownIdentification()}`
+        `The following dropdown has a placeholder option, but no placeholder text. Please use at least one of these props: placeholder, label, or helpText. ${identificationText}`
       )
 
       this.sentPlaceholderWarning = true
@@ -95,18 +94,19 @@ ${this.getDropdownIdentification()}`
       form,
       name,
       required,
-      inline,
+      variation,
     } = this.props
 
     const hasValidInitialValue =
       this.getOptionFromValue(this.initialValue) !== null
 
     const isPlaceholder = this.getSelectedOption() === null
+    const isInline = variation.toLowerCase() === 'inline'
     let width
     let iconSize
 
     let classes = 'bg-transparent bn w-100 h-100 '
-    let containerClasses = `${inline ? '' : 'bw1'} br2 relative `
+    let containerClasses = `${isInline ? '' : 'bw1'} br2 relative `
     let selectClasses = 'o-0 absolute top-0 left-0 w-100 bottom-00 '
 
     let labelClasses = 'vtex-dropdown__label db mb3 w-100 c-on-base '
@@ -114,7 +114,7 @@ ${this.getDropdownIdentification()}`
     const valueLabel = this.getValueLabel()
     const showCaption = !valueLabel
 
-    const color = inline ? 'c-link ' : 'c-on-base '
+    const color = isInline ? 'c-link ' : 'c-on-base '
 
     classes += disabled ? '' : 'pointer '
     selectClasses += disabled ? '' : 'pointer '
@@ -127,17 +127,17 @@ ${this.getDropdownIdentification()}`
 
     switch (size) {
       case 'small':
-        classes += inline ? 'ph2 ' : 'pl5 pr3 '
+        classes += isInline ? 'ph2 ' : 'pl5 pr3 '
         selectClasses += 't-small '
         labelClasses += 't-small '
-        containerClasses += `${inline ? 'h-auto' : 'h-small'} t-small `
+        containerClasses += `${isInline ? 'h-auto' : 'h-small'} t-small `
         iconSize = 18
         break
       case 'large':
-        classes += inline ? 'ph2 ' : 'ph5 '
+        classes += isInline ? 'ph2 ' : 'ph5 '
         selectClasses += 't-body '
         labelClasses += 't-body '
-        containerClasses += `${inline ? 'h-auto' : 'h-large'} t-body `
+        containerClasses += `${isInline ? 'h-auto' : 'h-large'} t-body `
         iconSize = 18
         break
       case 'x-large':
@@ -148,10 +148,10 @@ ${this.getDropdownIdentification()}`
         iconSize = 22
         break
       default:
-        classes += inline ? 'ph2 ' : 'pl5 pr4 '
+        classes += isInline ? 'ph2 ' : 'pl5 pr4 '
         selectClasses += 't-small '
         labelClasses += 't-small '
-        containerClasses += `${inline ? 'h-auto' : 'h-regular'} t-body `
+        containerClasses += `${isInline ? 'h-auto' : 'h-regular'} t-body `
         iconSize = 18
         break
     }
@@ -162,24 +162,25 @@ ${this.getDropdownIdentification()}`
       containerClasses += 'ba b--disabled bw1 bg-disabled '
     } else if (error || errorMessage) {
       containerClasses += 'ba b--danger hover-b--danger '
-    } else if (inline) {
+    } else if (isInline) {
       containerClasses += 'fw5 '
     } else {
       containerClasses += 'bg-base hover-b--muted-3 ba b--muted-4 '
     }
 
     return (
-      <div className={`vtex-dropdown ${inline ? 'dib' : ''}`}>
+      <div className={`vtex-dropdown ${isInline ? 'dib' : ''}`}>
         <label>
           {label && <span className={labelClasses}>{label}</span>}
           <div className={containerClasses} style={containerStyle}>
             <div id={id} className={`vtex-dropdown__button ${classes}`}>
-              <div className={`flex ${inline ? '' : 'h-100'}`}>
+              <div className={`flex ${isInline ? '' : 'h-100'}`}>
                 <div
                   className={`vtex-dropdown__caption flex-auto tl truncate ${
-                    inline ? '' : 'h-100'
+                    isInline ? '' : 'h-100'
                   }`}>
-                  <div className={`${inline ? '' : 'h-100'} flex items-center`}>
+                  <div
+                    className={`${isInline ? '' : 'h-100'} flex items-center`}>
                     {showCaption ? placeholder : valueLabel}
                   </div>
                 </div>
@@ -246,7 +247,7 @@ DropdownWithRef.displayName = 'Dropdown'
 DropdownWithRef.defaultProps = {
   size: 'regular',
   options: [],
-  inline: false,
+  variation: 'primary',
 }
 
 DropdownWithRef.propTypes = {
@@ -281,8 +282,8 @@ DropdownWithRef.propTypes = {
   value: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   /** Spec attribute */
   disabled: PropTypes.bool,
-  /** Spec attribute */
-  inline: PropTypes.bool,
+  /** Dropdown variation */
+  variation: PropTypes.oneOf(['primary', 'inline']),
   /** Spec attribute */
   form: PropTypes.string,
   /** Spec attribute */